remove API docs generation from the WILDDUCK specific tools.js file into a separate file. Do appropriate changes
